// giata-types.ts
// Generated by https://quicktype.io

export interface GiataListPropertiesResponse {
  properties: Properties;
}

export interface Properties {
  attributes: PropertiesAttributes;
  property: GiataProperty[];
  more: More;
}

export interface PropertiesAttributes {
  "xmlns:xsi": string;
  "xmlns:xlink": string;
  lastUpdate: string;
}

export interface More {
  attributes: MoreAttributes;
}

export interface MoreAttributes {
  "xlink:href": string;
}

export interface GiataProperty {
  attributes: PropertyAttributes;
  name: string;
  alternativeNames: AlternativeNames;
  city: City;
  destination: Destination;
  country: string;
  category: number;
  ratings: Ratings;
  airports: Airports;
  addresses: Addresses;
  phones: Phones;
  emails: Emails;
  urls: Urls;
  geoCodes: GeoCodes;
  propertyCodes: PropertyCodes;
  chains: Chains;
  ghgml: More;
}

export interface Addresses {
  address: Address;
}

export interface Address {
  addressLine: AddressLine[];
  street: string;
  cityName: string;
  country: string;
}

export interface AddressLine {
  text: string;
  attributes: AddressLineAttributes;
}

export interface AddressLineAttributes {
  addressLineNumber: string;
}

export interface Airports {
  airport: Airport[];
}

export interface Airport {
  attributes: AirportAttributes;
}

export interface AirportAttributes {
  iata: string;
}

export interface AlternativeNames {
  alternativeName: AlternativeName[];
}

export interface AlternativeName {
  text: string;
  attributes: AlternativeNameAttributes;
}

export interface AlternativeNameAttributes {
  alternativeNameType: string;
  effectiveDate?: string;
}

export interface PropertyAttributes {
  giataId: string;
  lastUpdate: string;
}

export interface Chains {
  chain: Chain;
}

export interface Chain {
  attributes: ChainAttributes;
}

export interface ChainAttributes {
  chainId: string;
  chainName: string;
}

export interface City {
  text: string;
  attributes: CityAttributes;
}

export interface CityAttributes {
  cityId: string;
}

export interface Destination {
  text: string;
  attributes: DestinationAttributes;
}

export interface DestinationAttributes {
  destinationId: string;
}

export interface Emails {
  email: string;
}

export interface GeoCodes {
  geoCode: GeoCode;
}

export interface GeoCode {
  attributes: GeoCodeAttributes;
  latitude: number;
  longitude: number;
}

export interface GeoCodeAttributes {
  accuracy: string;
}

export interface Phones {
  phone: Phone[];
}

export interface Phone {
  text: number;
  attributes: PhoneAttributes;
}

export interface PhoneAttributes {
  tech: string;
}

export interface PropertyCodes {
  provider: Provider[];
}

export interface Provider {
  attributes: ProviderAttributes;
  code: CodeElement[] | PurpleCode;
}

export interface ProviderAttributes {
  providerCode: string;
  providerType: ProviderType;
}

export enum ProviderType {
  Gds = "gds",
  TourOperator = "tourOperator"
}

export interface CodeElement {
  attributes?: CodeAttributes;
  value: PurpleValue[] | number | string;
}

export interface CodeAttributes {
  status: Status;
}

export enum Status {
  Inactive = "inactive"
}

export interface PurpleValue {
  text: number | string;
  attributes: ValueAttributes;
}

export interface ValueAttributes {
  name: string;
}

export interface PurpleCode {
  value: FluffyValue[] | number | string;
}

export interface FluffyValue {
  text: string;
  attributes: ValueAttributes;
}

export interface Ratings {
  rating: Rating;
}

export interface Rating {
  attributes: RatingAttributes;
}

export interface RatingAttributes {
  value: string;
  isDefault: string;
}

export interface Urls {
  url: string[];
}